+++ src/thing/thing.test.ts
@@ -154,7 +154,7 @@ describe("getThing", () => {
     const thing = getThing(
       datasetWithMultipleThings,
       "https://some.vocab/subject"
-    );
+    ) as Thing;
 
     expect(Array.from(thing)).toEqual([relevantQuad]);
   });
@@ -167,7 +167,7 @@ describe("getThing", () => {
     const thing = getThing(
       datasetWithAThing,
       DataFactory.namedNode("https://some.vocab/subject")
-    );
+    ) as Thing;
 
     expect(Array.from(thing)).toEqual([relevantQuad]);
   });
@@ -182,9 +182,23 @@ describe("getThing", () => {
     const datasetWithThingLocal = dataset();
     datasetWithThingLocal.add(quadWithLocalSubject);
 
-    const thing = getThing(datasetWithThingLocal, localSubject);
+    const thing = getThing(datasetWithThingLocal, localSubject) as Thing;
 
     expect(Array.from(thing)).toEqual([quadWithLocalSubject]);
+  });
+
+  it("returns null if the given SolidDataset does not include Quads with the given Subject", () => {
+    const datasetWithoutTheThing = dataset();
+    datasetWithoutTheThing.add(
+      getMockQuad({ subject: "https://arbitrary-other.vocab/subject" })
+    );
+
+    const thing = getThing(
+      datasetWithoutTheThing,
+      "https://some.vocab/subject"
+    );
+
+    expect(thing).toBeNull();
   });
 
   it("returns Quads from all Named Graphs if no scope was specified", () => {
@@ -202,7 +216,7 @@ describe("getThing", () => {
     const thing = getThing(
       datasetWithMultipleNamedGraphs,
       "https://some.vocab/subject"
-    );
+    ) as Thing;
 
     expect(thing.size).toEqual(2);
     expect(Array.from(thing)).toContain(quadInDefaultGraph);
@@ -227,7 +241,7 @@ describe("getThing", () => {
       datasetWithMultipleNamedGraphs,
       "https://some.vocab/subject",
       { scope: "https://some.vocab/namedGraph" }
-    );
+    ) as Thing;
 
     expect(Array.from(thing)).toEqual([relevantQuad]);
   });
@@ -250,7 +264,7 @@ describe("getThing", () => {
       datasetWithMultipleNamedGraphs,
       "https://some.vocab/subject",
       { scope: "https://some.vocab/namedGraph" }
-    );
+    ) as Thing;
 
     expect(Array.from(thing)).toEqual([relevantQuad]);
   });
@@ -273,7 +287,7 @@ describe("getThing", () => {
       datasetWithMultipleNamedGraphs,
       "https://some.vocab/subject",
       { scope: DataFactory.namedNode("https://some.vocab/namedGraph") }
-    );
+    ) as Thing;
 
     expect(Array.from(thing)).toEqual([relevantQuad]);
   });